[163] Use requester in comment and jobs controller
